Add more brutal nested closure test
As far as I understand, this is because the current mocked-up PAD lookup at
compile time just looks things up right there in the current(ly compiling)
PAD so it will not see whatever FAKEs would otherwise be found in the parent
PAD. Or something like that. It does not all line up just yet. :(