Implement Lua function calling/wrapping for Perl
Bit of a mammoth patch, sorry.

This implements generating blessed CODE refs in Perl from Lua. These are
blessed to be able to clean up the function reference within Lua from
Perl. On top of that, this implements calling those functions from Perl,
auto-converting parameters.

The necessary state of the function is attached to the CV using the
ALIAS/XSANY trick that was first introduced in Class::XSAccessor. For
details on how that works, please read the giant comment blocks in
the XSAccessor.xs file of that CPAN module.

For extended usage examples, see t/30_lua_functions.t. But in a
nutshell, this implements:

  my $func;
  lua {
    local f = function (a, b)
      return a + b
    end
    $func = f
  }

  say $func->(13, 15); # prints 28