Make lua_function use the right package
Also cf. new test in 13_embed_funcs.t.

lua_function appears to have been doing the wrong thing wrt. which
package it installs to. This now uses newXS's capability to install
named subs where appropriate. Memory management needed some twiddling
since the stash will want to hold a ref to the CV whereas the RV is no
longer useful.

This also does away with having to call into Perl for installing XSUBs.
That was a bit of a cop-out anyway.