Add support for building with lua 5.2.2
instead of luajit. There are some API compatibility macros
to be defined. The Makefile.PL tries to DRTR based on the
environment variables LUAJIT_DIR or LUA_DIR, but your
mileage may obviously vary.

NB. This patch is not enough to make all tests pass on
OSX, but at least lua is more friends with OSX than luajit.